requirements.txt
Telethon~=1.34.0
python-dotenv~=1.0.1
pytz~=2024.1
requests~=2.24.0